test(reborn): isolate live QA semantic judge
